requirements.txt
django~=2.1
social-auth-app-django~=3.1
python-jose~=3.0
python-dotenv~=0.9

Pyvimeo~=1.1.0
